ct_dashboard_api: extract auth business logic to api layer
